67314155 - test

+++ src/main/scala/job/remote/DistinctUserEmbeddingJob.scala
@@ -7,7 +7,7 @@ object DistinctUserEmbeddingJob extends RemoteSparkJob{
         val path = "viewfs://hadoop-meituan/user/hadoop-hmart-waimaiad/yangyufeng04/bigmodel/afo_model_newpage_life/20220718_190358/"
         val file = sc.textFile(path + "user_embedding")
         val data = file.map(row => {
-            val expose_time = row.split(',')(0).toLong
+            val expose_time = row.split(',')(0).toDouble.toLong
             val uuid = row.split(',')(1)
             val user_emb = row.split(',').slice(2, row.length).map(_.toDouble)
             (uuid, (expose_time, user_emb))
